requirements.txt
annotated-types==0.6.0
anyio==4.3.0
autopep8==2.1.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
